+++ src/Hledger/Flow/BaseDir.hs
@@ -1,7 +1,13 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE QuasiQuotes #-}
 
-module Hledger.Flow.BaseDir where
+module Hledger.Flow.BaseDir (
+    determineBaseDir
+  , relativeToBase
+  , relativeToBase'
+  , turtleBaseDir
+  , effectiveRunDir
+) where
 
 import Path
 import Path.IO
@@ -12,9 +18,11 @@ import Data.Maybe
 
 import Control.Monad.Catch (MonadThrow, throwM)
 import Control.Monad.IO.Class (MonadIO)
+import Control.Monad (when)
 
-
-import qualified Turtle as Turtle (stripPrefix)
+import qualified Turtle (liftIO, repr, stripPrefix)
+import qualified Data.Text as T
+import qualified Data.Text.IO as T
 
 determineBaseDir :: Maybe TurtlePath -> IO (BaseDir, RunDir)
 determineBaseDir suppliedDir = do
@@ -32,13 +40,33 @@ determineBaseDirFromStartDir startDir = determineBaseDirFromStartDir' startDir s
 
 determineBaseDirFromStartDir' :: (MonadIO m, MonadThrow m) => AbsDir -> AbsDir -> m (BaseDir, RunDir)
 determineBaseDirFromStartDir' startDir possibleBaseDir = do
-  _ <- if (parent possibleBaseDir == possibleBaseDir) then throwM (MissingBaseDir startDir) else return ()
+  Control.Monad.when (parent possibleBaseDir == possibleBaseDir) $ throwM (MissingBaseDir startDir)
   foundBaseDir <- doesDirExist $ possibleBaseDir </> [reldir|import|]
   if foundBaseDir then
     do
-      runDir <- makeRelative possibleBaseDir startDir
+      runDir <- limitRunDir possibleBaseDir startDir
       return (possibleBaseDir, runDir)
     else determineBaseDirFromStartDir' startDir $ parent possibleBaseDir
+
+-- | We have unexpected behaviour when the runDir is deeper than the account directory,
+-- e.g. "1-in" or the year directory. Specifically, include files are generated incorrectly
+-- and some journals are written entirely outside of the baseDir.
+-- limitRunDir can possibly removed if the above is fixed.
+limitRunDir :: (MonadIO m, MonadThrow m) => BaseDir -> AbsDir -> m RunDir
+limitRunDir bd absRunDir = do
+  rel <- makeRelative bd absRunDir
+  let runDirDepth = pathSize rel
+  let fun = composeN (runDirDepth - 4) parent
+  let newRunDir = fun rel
+  when (runDirDepth > 4) $ do
+    let msg = T.pack $ "Changing runDir from " ++ Turtle.repr rel ++ " to " ++ Turtle.repr newRunDir :: T.Text
+    Turtle.liftIO $ T.putStrLn msg
+  return newRunDir
+
+composeN :: Int -> (a -> a) -> (a -> a)
+composeN n f | n < 1      = id
+             | n == 1     = f
+             | otherwise = composeN (n-1) (f . f)
 
 relativeToBase :: HasBaseDir o => o -> TurtlePath -> TurtlePath
 relativeToBase opts = relativeToBase' $ pathToTurtle (baseDir opts)
@@ -50,10 +78,8 @@ relativeToBase' bd p = if forceTrailingSlash bd == forceTrailingSlash p then "./
 turtleBaseDir :: HasBaseDir o => o -> TurtlePath
 turtleBaseDir opts = pathToTurtle $ baseDir opts
 
-effectiveRunDir :: BaseDir -> RunDir -> Bool -> AbsDir
-effectiveRunDir bd rd useRunDir = do
+effectiveRunDir :: BaseDir -> RunDir -> AbsDir
+effectiveRunDir bd rd = do
   let baseImportDir = bd </> [Path.reldir|import|]
   let absRunDir = bd </> rd
-  if useRunDir
-    then if absRunDir == bd then baseImportDir else absRunDir
-    else baseImportDir
+  if absRunDir == bd then baseImportDir else absRunDir